Rearrange new file and add new feature
This commit is arranging file for features directory and addi new feature
for facial recognition using camera snapshot. and it is work and tested properly
as long as:

1. The lighting is not darker that could cause face to be undetected
2. For a 1 meter distance also already tested work. and it is also work for fail
authentication if the image that is analyzed is different (different person).
3. I've tried it to auth using my selfie from phone by put my phone screen to
webcam and it does not work which is correctly as we expect.